requirements.txt
google-generativeai
pynput
Pillow
pyscreenshot
pydantic
pydantic-settings
pytesseract 